+++ include/seqan/blast.h
@@ -65,10 +65,9 @@
 #include "blast/blast_tabular.h"
 #include "blast/blast_io_context.h"
 #include "blast/blast_statistics.h"
-#include "blast/blast_tabular_read.h"
-#include "blast/blast_tabular_write.h"
-#include "blast/blast_tabular_formatted_file.h"
-#include "blast/blast_report_write.h"
+#include "blast/blast_tabular_in.h"
+#include "blast/blast_tabular_out.h"
+#include "blast/blast_report_out.h"
 
 #else //SEQAN_C++11_STANDARD
